[SHK-280] Member 코드 구조 개선 (#60)
* Refactor : [SHK-286] url 컨벤션에 맞게 수정

* Refactor : [SHK-287] deleteDeliveryInfo 누락된 @Transactional 추가

* Refactor : [SHK-288] validateAccess 메서드화

* Refactor : [SHK-289] getMemberEntity 메서드화

* Refactor : [SHK-290] getDeliveryEntity 메서드화

* Refactor : [SHK-291] bean validation 누락된 message 작성

* Refactor : updateMember id -> memberId로 수정

+++ src/main/java/com/prgrms/kream/domain/member/service/MemberService.java
@@ -4,7 +4,6 @@ import static com.prgrms.kream.common.jwt.JwtUtil.*;
 import static com.prgrms.kream.common.mapper.MemberMapper.*;
 
 import java.util.List;
-import java.util.Objects;
 
 import javax.persistence.EntityNotFoundException;
 
@@ -29,11 +28,11 @@ import com.prgrms.kream.domain.member.dto.request.MemberUpdateServiceRequest;
 import com.prgrms.kream.domain.member.dto.response.DeliveryInfoGetResponse;
 import com.prgrms.kream.domain.member.dto.response.DeliveryInfoRegisterResponse;
 import com.prgrms.kream.domain.member.dto.response.DeliveryInfoUpdateResponse;
+import com.prgrms.kream.domain.member.dto.response.FollowingGetAllResponse;
 import com.prgrms.kream.domain.member.dto.response.MemberGetFacadeResponse;
 import com.prgrms.kream.domain.member.dto.response.MemberLoginResponse;
 import com.prgrms.kream.domain.member.dto.response.MemberRegisterResponse;
 import com.prgrms.kream.domain.member.dto.response.MemberUpdateServiceResponse;
-import com.prgrms.kream.domain.member.dto.response.FollowingGetAllResponse;
 import com.prgrms.kream.domain.member.model.DeliveryInfo;
 import com.prgrms.kream.domain.member.model.Following;
 import com.prgrms.kream.domain.member.model.FollowingId;
@@ -85,24 +84,16 @@ public class MemberService {
 
 	@Transactional(readOnly = true)
 	public MemberGetFacadeResponse getMember(Long id) {
-		if (!Objects.equals(getMemberId(), id)) {
-			throw new AccessDeniedException("잘못된 접근입니다");
-		}
-
-		Member member = memberRepository.findById(id)
-				.orElseThrow(() -> new EntityNotFoundException("존재하지 않은 회원입니다."));
+		validateAccess(id);
+		Member member = getMemberEntity(id);
 		return toMemberGetFacadeResponse(member);
 	}
 
 	@Transactional
 	public MemberUpdateServiceResponse updateMember(MemberUpdateServiceRequest memberUpdateServiceRequest) {
-		if (!isValidAccess(memberUpdateServiceRequest.id())) {
-			throw new AccessDeniedException("잘못된 접근입니다");
-		}
+		validateAccess(memberUpdateServiceRequest.id());
 
-		Member member = memberRepository.findById(memberUpdateServiceRequest.id())
-				.orElseThrow(() -> new EntityNotFoundException("존재하지 않는 회원입니다."));
-
+		Member member = getMemberEntity(memberUpdateServiceRequest.id());
 		member.updateMember(
 				memberUpdateServiceRequest.name(),
 				memberUpdateServiceRequest.phone(),
@@ -114,9 +105,7 @@ public class MemberService {
 
 	@Transactional(readOnly = true)
 	public Page<DeliveryInfoGetResponse> getDeliveryInfoPage(Long memberId, Pageable pageable) {
-		if (!isValidAccess(memberId)) {
-			throw new AccessDeniedException("잘못된 접근입니다.");
-		}
+		validateAccess(memberId);
 
 		return deliveryInfoRepository.findAllByMemberId(memberId, pageable)
 				.map(MemberMapper::toDeliveryInfoGetResponse);
@@ -124,21 +113,16 @@ public class MemberService {
 
 	@Transactional
 	public DeliveryInfoRegisterResponse registerDeliveryInfo(DeliveryInfoRegisterRequest deliveryInfoRegisterRequest) {
-		if (!isValidAccess(deliveryInfoRegisterRequest.memberId())) {
-			throw new AccessDeniedException("잘못된 접근입니다.");
-		}
+		validateAccess(deliveryInfoRegisterRequest.memberId());
+
 		DeliveryInfo deliveryInfo = deliveryInfoRepository.save(toDeliveryInfo(deliveryInfoRegisterRequest));
 		return new DeliveryInfoRegisterResponse(deliveryInfo.getId());
 	}
 
 	@Transactional
 	public DeliveryInfoUpdateResponse updateDeliveryInfo(DeliveryInfoUpdateRequest deliveryInfoUpdateRequest) {
-		if (!isValidAccess(deliveryInfoUpdateRequest.memberId())) {
-			throw new AccessDeniedException("잘못된 접근입니다.");
-		}
-
-		DeliveryInfo deliveryInfo = deliveryInfoRepository.findById(deliveryInfoUpdateRequest.deliveryInfoId())
-				.orElseThrow(() -> new EntityNotFoundException("존재하지 않는 배송정보입니다."));
+		validateAccess(deliveryInfoUpdateRequest.memberId());
+		DeliveryInfo deliveryInfo = getDeliveryInfoEntity(deliveryInfoUpdateRequest.deliveryInfoId());
 
 		deliveryInfo.updateDeliveryInfo(
 				deliveryInfoUpdateRequest.name(),
@@ -151,7 +135,10 @@ public class MemberService {
 		return MemberMapper.toDeliveryInfoUpdateResponse(deliveryInfo);
 	}
 
+	@Transactional
 	public void deleteDeliveryInfo(DeliveryInfoDeleteRequest deliveryInfoDeleteRequest) {
+		DeliveryInfo deliveryInfo = getDeliveryInfoEntity(deliveryInfoDeleteRequest.deliveryInfoId());
+		validateAccess(deliveryInfo.getMemberId());
 		deliveryInfoRepository.deleteById(deliveryInfoDeleteRequest.deliveryInfoId());
 	}
 
@@ -191,5 +178,21 @@ public class MemberService {
 				.toList();
 
 		return new FollowingGetAllResponse(followedMemberIds);
+	}
+
+	private void validateAccess(Long id) {
+		if (!isValidAccess(id)) {
+			throw new AccessDeniedException("잘못된 접근입니다");
+		}
+	}
+
+	private Member getMemberEntity(Long id) {
+		return memberRepository.findById(id)
+				.orElseThrow(() -> new EntityNotFoundException("존재하지 않은 회원입니다."));
+	}
+
+	private DeliveryInfo getDeliveryInfoEntity(Long id) {
+		return deliveryInfoRepository.findById(id)
+				.orElseThrow(() -> new EntityNotFoundException("존재하지 않는 배송정보입니다."));
 	}
 }
